refactor(ports): extract etb_replacement + tokens modules (PR #52 H1 fix)
Address the H1 finding from the /code-review on PR #52: ``ports.py``
had grown to 916 lines (>800-line guideline). Extracts the two
unrelated subsystems that contributed the bulk of the file's recent
growth:

- ``src/mtg_synergy_graph/etb_replacement.py`` (104 lines): the
  `_parse_etb_replacement_keyword` + `extract_etb_replacement_ports`
  helpers added by this PR. Lazy-imports `extract_effect_ports` from
  ``ports`` to avoid a top-level cycle (same pattern as
  ``copy_face_from.py`` after PR #50's cleanup).
- ``src/mtg_synergy_graph/tokens.py`` (108 lines): the pre-existing
  `_TOKEN_COLOR_MAP` + `_parse_token_script` + `_colors_for_prefix`
  helpers. Pure migration — no behaviour change.

``ports.py`` drops 916 → 762 lines, comfortably under the guideline.
Re-exports both modules' public symbols so existing call sites
(``importer.py``, ``tests/test_importer.py``, ``tests/test_ports.py``,
the public ``__init__.py`` surface) continue to work unchanged.

Pure refactor — `bench.py audit --expect-identity` PASS
(`tests/fixtures/golden_set_run.json`). 1879 tests pass at 90.51%
coverage.
